+++ src/lib/input-validation/validateComplexPassword.ts
@@ -11,11 +11,18 @@ import { ErrorTypes } from "./constant/errorTypes";
 export const ValidateComplexPasswordErrorType = Enum(ErrorTypes.empty, ErrorTypes.invalidFormat);
 export type ValidateComplexPasswordErrorType = Enum<typeof ValidateComplexPasswordErrorType>;
 
-export const validateComplexPassword = (value: string) => {
+export interface ValidateComplexPasswordOptions {
+  isRequired?: boolean;
+}
+
+export const validateComplexPassword = (value: string, { isRequired = true }: ValidateComplexPasswordOptions = {}) => {
   const REGEX =
     // tslint:disable-next-line
     '^(?:(?=.*\\d)(?=.*[A-Z])(?=.*[a-z])|(?=.*\\d)(?=.*[^A-Za-z0-9])(?=.*[a-z])|(?=.*[^A-Za-z0-9])(?=.*[A-Z])(?=.*[a-z])|(?=.*\\d)(?=.*[A-Z])(?=.*[^A-Za-z0-9]))(?!.*(.)\\1{2,})[A-Za-z0-9!~<>,;:_=?*+#."&§%°()\\|\\[\\]\\-\\$\\^\\@\\/]{8,32}$';
   if (isEmpty(value)) {
+    if (!isRequired) {
+      return null;
+    }
     return ValidateComplexPasswordErrorType.empty;
   }
   if (!matches(value, REGEX)) {