+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- Adding isRequired option for some validate function
 - Adding isAllValid method, to check whether all value in validation is valid or not
 - Added Validation class
 - Added validateAlphanumeric
@@ -17,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added validateEmail
 - Added validateLength
 - Added validateMemorySize
+- Added validateNotEmpty
 - Added validateNumeric
 - Added validateOrderNumber
 - Added validatePath